SSH Bastion Support for initial provision/rke
Helm isn't covered by this. We could either setup tunnel or make
sure helm is on the remote bastion and just script helm remotely
on the bastion in those circumtances, however, I'm not planning
on going the last mile with this.

I'm switching gears to doing a remote-exec in the terraform
module. It means there will be a couple extra steps for some sort
of future proxy/limited internet support, but the overall execution
is simpler.

I'm still going to put this into a PR and partially revert it,
since perhaps the easyssh module is still valueable. Even if not,
we'll have this for posterity if we need to support bastions in
RanchHand directly at some point.