Update tool and chart versions (#8)
* update local tool versions

updating to the latest versions of kubectl, helm and rke

* updates rancher helm chart version

rancher changed its versioning scheme so that the chart matches the
version of rancher itself. the old version is no longer
available...rude!

* logs to stdout instead of stderr

for some reason logrus defaults to logging to stderr (just like scumbag
npm). this change moves log output to stdout