CA cert usage adjustments (#7)
* Use our CA as private ca in rancher

I was wrong--there is an extra level of validation here.

* Rancher agent does care about IPs, switch to hosts

* additionalTrustedCAs support

Note: I think this can be reverted, but commiting this just in case
since I haven't proven it.

* But kubectl requires this. Sigh.

* Configureable cert sans/ips

* Add default arg for cert vals, update help

* Remove unnecessary empty assignment

* migrates ca cert generation out of helm pkg

tls cert is generated and persisted to output directory so subsequent
runs are idempotent. push rancher cacert secert creation logic into
rancher.go file. renames installKubernetes() to launchRKE(); easier to
navigate filenames

* logs when docker install begins

because this takes a while and we should indicate why execution has
appeared to pause

* Dumb conditional tweak

Since this is an actual logic branch rather than a guard block, I
thought it would be better in an else block.